refactor(core): break grants/skills/races import cycles
Extract PHB_SKILL_IDS to skill_ids leaf, move build_fixed_proficiencies
to grants_resolve only, hoist cluster lazy imports, and keep
finalize_creation in _creation_finalize.
