+++ core/skill_ids.py
@@ -1,0 +1,7 @@
+"""Id навыков PHB — leaf без импортов races/grants/skills."""
+
+from core.abilities import skill_ids
+
+PHB_SKILL_IDS: tuple[str, ...] = skill_ids()
+
+THIEVES_TOOLS_ID = "thieves_tools"
