+++ ui/menus/_creation_steps.py
@@ -4,27 +4,10 @@ from core.localization import get_string
 from core.models import Character
 from core.types import StringsDict
 from ui.input_handler import get_str_input
-from ui.menus._creation_finalize import (
-    save_created_character,
-)
-from ui.menus._creation_state import _CreationState
-from ui.menus.console import (
-    print_screen_header,
-    print_success_and_wait,
-)
+from ui.menus._creation_handlers import _STEP_HANDLERS
+from ui.menus._creation_state import CreationStep, _CreationState
+from ui.menus.console import print_screen_header
 from ui.menus.settings import select_difficulty
-
-
-def finalize_creation(
-    strings: StringsDict, state: _CreationState
-) -> Character | None:
-    """Сохранить персонажа и показать сообщение об успехе."""
-    character = save_created_character(state)
-    if character is None:
-        return None
-    msg = get_string(strings, "character.save_success", name=state.name)
-    print_success_and_wait(strings, msg)
-    return character
 
 
 def show_create_character_flow(
@@ -54,9 +37,6 @@ def run_creation_steps(
     language: str = "ru",
 ) -> Character | None:
     """Цикл шагов создания персонажа после ввода имени."""
-    from ui.menus._creation_handlers import _STEP_HANDLERS
-    from ui.menus._creation_state import CreationStep
-
     step: CreationStep = "race"
 
     while True:
